info_portal_services/requirements.txt
Django==4.0.3
gunicorn==20.1.0
psycopg2-binary==2.9.3
click==8.1.0
djangorestframework==3.13.1
PyJWT>=2.3.0
requests==2.27.1
django_cors_headers==3.11.0
drf-yasg==1.20.0
Pillow==9.0.1
pydantic==1.9.0
mozilla-django-oidc==2.0.0
drf-yasg==1.20.0
django-cors-headers==3.11.0
djangorestframework-gis==0.18
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.9
